% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treezoom.R
\name{treezoom}
\alias{treezoom}
\title{'d3 v4 Zooomable Treemap htmlwidget'}
\usage{
treezoom(data = NULL, sizeField = "value", valueField = NULL,
  labelField = "name", style = list(fill = "none", stroke = "black"),
  styleRect = NULL, styleText = NULL, clipText = FALSE,
  margin = list(left = 0, top = 0, right = 0, bottom = 0), ...,
  width = "100\%", height = NULL, elementId = NULL)
}
\description{
Create a d3 v4 based treemap htmlwidget
 as a foundation for other interactivity
}
