% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree.R
\name{tree}
\alias{tree}
\title{'d3 v4 Treemap htmlwidget'}
\usage{
tree(data = NULL, sizeField = "value", valueField = NULL,
  labelField = "name", style = list(fill = "none", stroke = "black"),
  styleRect = styleRect, styleText = styleText, ..., width = "100\%",
  height = NULL, elementId = NULL)
}
\description{
Create a d3 v4 based treemap htmlwidget
 as a foundation for other interactivity
}
